requirements.txt
openai==0.28
gradio
PyPDF2
python-docx
python-dotenv